+++ UnityProject/Assets/Scripts/US13/Player/HUDData/HUDHandler.cs
@@ -44,7 +44,5 @@ namespace US13.Player.HUDData
 				Categorys[type].Remove(HUD);
 			}
 		}
-
-
 	}
 }
